140-dev-chain-of-asyn-derivation: added PendingUpdate status; modified updateDerivationAsyn method to cope with chain of asynchronous derivation when accessing information.

+++ JPS_BASE_LIB/src/main/java/uk/ac/cam/cares/jps/base/derivation/DerivationClient.java
@@ -203,7 +203,8 @@ public class DerivationClient {
 		// the graph object makes sure that there is no circular dependency
 		DirectedAcyclicGraph<String,DefaultEdge> graph = new DirectedAcyclicGraph<String,DefaultEdge>(DefaultEdge.class);
 		try {
-			updateDerivationAsyn(derivationIRI, graph);
+			// the flag upstreamDerivationRequested is set as false by default
+			updateDerivationAsyn(derivationIRI, graph, false);
 		} catch (Exception e) {
 			LOGGER.fatal(e.getMessage());
 			throw new JPSRuntimeException(e);
@@ -351,6 +352,15 @@ public class DerivationClient {
 	public String getDerivationOf(String entity) {
 		return this.sparqlClient.getDerivedIRI(entity);
 	}
+	
+	/**
+	 * Checks if the derivation status is "PendingUpdate".
+	 * @param derivation
+	 * @return
+	 */
+	public boolean isPendingUpdate(String derivation) {
+		return this.sparqlClient.isPendingUpdate(derivation);
+	}
 
 	/**
 	 * Checks if the derivation status is "Requested".
@@ -379,6 +389,14 @@ public class DerivationClient {
 		return this.sparqlClient.isFinished(derivation);
 	}
 
+	/**
+	 * Marks the derivation status as "PendingUpdate".
+	 * @param derivation
+	 */
+	public void markAsPendingUpdate(String derivation) {
+		this.sparqlClient.markAsPendingUpdate(derivation);
+	}
+	
 	/**
 	 * Marks the derivation status as "Requested".
 	 * @param derivation
@@ -457,7 +475,7 @@ public class DerivationClient {
 	 * @param instance
 	 * @param graph
 	 */
-	private void updateDerivationAsyn(String instance, DirectedAcyclicGraph<String, DefaultEdge> graph) {
+	private void updateDerivationAsyn(String instance, DirectedAcyclicGraph<String, DefaultEdge> graph, boolean upstreamDerivationRequested) {
 		// this method follows the first a few steps of method updateDerivation(String instance, DirectedAcyclicGraph<String, DefaultEdge> graph)
 		// TODO in future development, ideally these two method should be merged into the same method?
 		List<String> inputsAndDerived = this.sparqlClient.getInputsAndDerived(instance);
@@ -471,25 +489,47 @@ public class DerivationClient {
 				graph.addVertex(input);
 			}
 			graph.addEdge(instance, input);
-			updateDerivationAsyn(input, graph);
+			updateDerivationAsyn(input, graph, upstreamDerivationRequested);
 		}
 		
 		List<String> inputs = this.sparqlClient.getInputs(instance);
 		if (inputs.size() > 0) {
-			if (isOutOfDate(instance,inputs)) {
-				// if the Derivation is out of date, the first thing we do is checking its status
-				if (!this.sparqlClient.hasStatus(instance)) {
-					// if there's no status, then mark as requested - a job need to be started to update the derivation
-					LOGGER.info("Updating <" + instance + ">, marked as Requested");
-					LOGGER.debug("<" + instance + "> is out-of-date when compared to <" + inputs + ">");
-					this.sparqlClient.markAsRequested(instance);
+			// here only derivation instance will enter, first we check if it is an asynchronous derivation
+			if (isDerivedAsynchronous(instance)) {
+				// then we check if this derivation is part of a chain and if any upstream one is already requested
+				// as the upstreamDerivationRequested flag is false by default, the code will directly go to the else part
+				if (upstreamDerivationRequested) {
+					// here it means one of the upstream derivation is already requested
+					// so we check if the current derivation has status already
+					// if there is status already, the derivation framework just pass
+					// otherwise, we mark it as PendingUpdate
+					if (!this.sparqlClient.hasStatus(instance)) {
+						LOGGER.info("Pending to update <" + instance + ">, marked as PendingUpdate");
+						this.sparqlClient.markAsPendingUpdate(instance);
+					}
 				} else {
-					// for now, if there's any status, the derivation framework just pass
-				}
-			} else {
-				// if the derivation is up to date, then delete <hasStatus> <Status> if applies
-				if (this.sparqlClient.hasStatus(instance)) {
-					this.sparqlClient.deleteStatus(instance);
+					// only if all upstream ones are up-to-date, we check if the current derivation is out-of-date
+					// this applies to the first derivation in the chain by default
+					if (isOutOfDate(instance,inputs)) {
+						// if the Derivation is out of date, the first thing we do is checking its status
+						if (!this.sparqlClient.hasStatus(instance)) {
+							// if there's no status, then mark as requested - a job need to be started to update the derivation
+							LOGGER.info("Updating <" + instance + ">, marked as Requested");
+							LOGGER.debug("<" + instance + "> is out-of-date when compared to <" + inputs + ">");
+							this.sparqlClient.markAsRequested(instance);
+						} else {
+							// for now, if there's any status, the derivation framework just pass
+						}
+						// as at this point, this should be the first derivation in the chain to be updated
+						// we set the flag to remind all downstream derivations
+						upstreamDerivationRequested = true;
+					} else {
+						// if the derivation is up to date, then delete <hasStatus> <Status> if applies
+						// as in theory, here flag upstreamDerivationRequested is false, so this derivation remains up-to-date
+						if (this.sparqlClient.hasStatus(instance)) {
+							this.sparqlClient.deleteStatus(instance);
+						}
+					}
 				}
 			}
 		}
